requirements/development.txt
-r base.txt
coverage==4.0.3
django-extensions==1.6.1
django-debug-toolbar==1.4
flake8==2.5.4
pygraphviz==1.3.1
